==> part1-team-data-retrieval.ipynb <==
## NCAA March Madness Prediction (Pt. 1)

import numpy as np
import pandas as pd

import warnings
warnings.filterwarnings("ignore")

#### Part 1: Data Retrieval
Purpose: Given Raw Data of all Division 1 Basketball Schools from the 2017-2018 season, retrieve all NCAA March Madness participants.

# Loading in Raw Data of School Features
basic = pd.read_csv("data/season1718/basic_school.csv")
opponents = pd.read_csv("data/season1718/basic_opps.csv")
advanced = pd.read_csv("data/season1718/adv_school.csv")
opp_adv = pd.read_csv("data/season1718/adv_opps.csv")

print(basic.shape)
basic.head(3)

# NOTE: Do not run this block to generate data for all teams

# Remove all schools that are not March Madness Tourney participants
basic = basic[(basic["School"].str.contains("NCAA"))]
opponents = opponents[(opponents["School"].str.contains("NCAA"))]
advanced = advanced[(advanced["School"].str.contains("NCAA"))]
opp_adv = opp_adv[(opp_adv["School"].str.contains("NCAA"))]

# Check the above (68 Total Competitors, Same Schools in every Table)
assert basic.shape[0] == 68
assert opponents.shape[0] == 68
assert advanced.shape[0] == 68
assert opp_adv.shape[0] == 68

assert sorted(basic["School"].values) == sorted(opponents["School"].values)
assert sorted(opponents["School"].values) == sorted(advanced["School"].values)
assert sorted(opponents["School"].values) == sorted(opp_adv["School"].values)

# Remove "NCAA" String from School Names
# *WARNING* - Do not run previous block after this one, otherwise will eliminate all rows
def remove_NCAA_str(row):
    return row["School"].replace("NCAA", "").strip()

basic["School"] = basic.apply(remove_NCAA_str, axis=1)
opponents["School"] = opponents.apply(remove_NCAA_str, axis=1)
advanced["School"] = advanced.apply(remove_NCAA_str, axis=1)
opp_adv["School"] = opp_adv.apply(remove_NCAA_str, axis=1)

#### Data Cleaning
Purpose: Begin to Remove Columns that are unnecessary before merge

# Basic Table:
# Dropped: [Rk, G, W-L%, Unnamed: 16, FG%, FT%, 3P%, FT%]
# Reminders: SRS - Simple Rating System, SOS - Strength of Schedule
# W,L.1 - Conference, W,L.2 - Home, W,L.3 - Away
# Tm./Opp. - Points for/against
redundant = ['Rk', 'G', 'W-L%', 'Unnamed: 16', 'FG%', 'FT%', '3P%', 'FT%']
try:
    basic.drop(redundant, axis=1, inplace=True)
except ValueError as err:
    print("Columns Already Removed from Basic Table")

basic.reset_index(drop=True, inplace=True)
basic.head()

# Opponents Table
keep = ["School", "FG", "FGA", "3P", "3PA", "FT", "FTA", "ORB", "TRB", "AST", "STL", "BLK", "TOV", "PF"]
try:
    opponents.drop(opponents.columns.difference(keep), axis=1, inplace=True)
except ValueError as err:
    print("Columns already removed from Opponents Table")

# Renaming Columns before Merging with 
opponents.reset_index(drop=True, inplace=True)
opponents.head()

# Advanced Statistics Table
original_index = basic.columns.append(pd.Index(redundant))
common_cols = original_index.intersection(advanced.columns)
to_remove = common_cols.drop(["School"])
# print("Dropped: ", to_remove.tolist())

try:
    advanced.drop(to_remove, axis=1, inplace=True)
except ValueError as err:
    print("Columns already removed from Advanced Table")

advanced.reset_index(drop=True, inplace=True)
advanced.head()

# Advanced Statistics (Opponents) Table
common_cols = original_index.intersection(opp_adv.columns)
to_remove = common_cols.drop(["School"])

try:
    opp_adv.drop(to_remove, axis=1, inplace=True)
except ValueError as err:
    print("Columns already removed from Advanced (Opponents) Table")

opp_adv.reset_index(drop=True, inplace=True)
opp_adv.head()

#### Merge Tables
Purpose: Construct an extensive feature vector of each school

# Combine features for each school from 4 separate DF's into one dataframe
opponents.columns = [keep[0]] + ["Opp_" + x for x in keep[1:]]
print("Opponent Columns Renamed: ", opponents.columns.tolist())
combined = basic.merge(opponents, how="inner", on="School")
# assert combined.shape[0] == 68 # Comment out for non-tourney participants
assert combined.shape[1] == (basic.shape[1]+opponents.shape[1])-1 # -1 for "School"

if "Opp_3PAr" not in opp_adv.columns.tolist():
    orig = opp_adv.columns.tolist()
    opp_adv.columns = [orig[0]] + ["Opp_" + x for x in orig[1:]]
print("Advanced (Opponents) Columns Renamed: ", opp_adv.columns.tolist())
combined2 = advanced.merge(opp_adv, how="inner", on="School")
# assert combined2.shape[0] == 68 # Comment out for non-tourney participants
assert combined2.shape[1] == (advanced.shape[1]+opp_adv.shape[1])-1

# Ensure only overlapping column is school
assert combined2.columns.intersection(combined.columns) == pd.Index(["School"])

combined3 = combined.merge(combined2, how="inner", on="School")
# assert combined3.shape[0] == 68 # Comment out for non-tourney participants
assert combined3.shape[1] == (combined.shape[1]+combined2.shape[1])-1

combined3.head()

#### Saving Results
Purpose: Transfer dataframe values to CSV to be used in next notebook for EDA analysis.
Modularizing the code means chunks of code don't have to be run repeatedly.

combined3.to_csv("data/cleaned_teams_1718.csv", index=False)
print(combined3.columns)
combined3.head()
