==> part2-game-data-retrieval.ipynb <==
## NCAA March Madness Prediction (Pt. 2)

import numpy as np
import pandas as pd
import random
import time
from colorama import Fore

import warnings
warnings.filterwarnings("ignore")

#### Part 1: Importing Teams + Games Data

cleaned_teams = pd.read_csv("data/cleaned_teams_1718.csv")
game_data = pd.read_csv("data/games.csv", dtype={'w_l': str}) 

#### Part 2: Cleaning Game Data (Extract 2017-2018 Data)

# Extracting games between 2017-2018 Season, Regular Season, before NCAA Tournament
start_date = "2017-11-10"
end_date = "2018-03-11"
games_1718 = game_data[(game_data["date"] >= start_date) & (game_data["date"] <= end_date)]

# March Madness Tournament Games
games_1718_tourney = game_data[(game_data["date"] > end_date)]

games_1718.reset_index(inplace=True)

# Dropping all columns except for teams and outcome
# This data would not be known before the game began, so it is not relevant
games_1718.drop(games_1718.columns.difference(["team_name", "opp_name", "w_l"]), axis=1, inplace=True)
print(games_1718.shape)
games_1718.head(3)

**Side Note**: Code for creating a map for school names (Purpose: bridge differences between datasets)

team_name_map = game_data[["team_name"]].drop_duplicates("team_name")
team_name_map.reset_index(inplace=True, drop=True)
# Removing Mascot from team name map
team_name_map["Name 2"] = team_name_map.apply(lambda x: " ".join(x["team_name"].split()[:-1]), axis=1)
team_name_map.columns = ["Name 1", "Name 2"]
# team_name_map.to_csv("data/team_name_map.csv", index=False) # Commented out to prevent overwriting

team_name_map = pd.read_csv("data/team_name_map.csv")
team_name_map.head(3)

#### Part 3: Create Feature Vector based on teams

# NOTE: NO LONGER NECESSARY
# Sampling originally for debugging purposes, trying the below cells on a smaller subset of the table

# Taking Random Sample of Games Data
sample_size = 0.01 * int(games_1718.shape[0])
print("Sample Size:", sample_size, "games")
sample = games_1718.sample(n=sample_size, axis=0, )
sample.reset_index(drop=True, inplace=True)
sample.head()

# Function for retrieving corresponding team names
def map_team_name(name):
    if name in team_name_map["Name 1"].values:
        return team_name_map[team_name_map["Name 1"] == name]["Name 2"].values[0]
    return name

# Define new feature vector data frames
feature_cols = cleaned_teams.columns.difference(["School"])
games_1718_x = pd.DataFrame(columns=feature_cols)
games_1718_y = pd.DataFrame(columns=["w_l"])

# Create vector quantifying difference between team's stats
second_call = False
def create_vector(row):
    global games_1718_x
    global games_1718_y
    global second_call
    
    # Perform string matching to compensate for 
    home_team = map_team_name(row["team_name"])
    away_team = map_team_name(row["opp_name"])
    
    home = cleaned_teams[cleaned_teams["School"] == home_team].drop("School", axis=1)
    away = cleaned_teams[cleaned_teams["School"] == away_team].drop("School", axis=1)
    diff_vector = home.values - away.values
    new_row = pd.DataFrame(data=diff_vector, columns=feature_cols)
    
#     print(home_team, "vs.", away_team)
    # print(diff_vector.shape)
    # print("Difference: ", new_row)
    
    games_1718_x = games_1718_x.append(new_row)
    games_1718_y = games_1718_y.append(row[["w_l"]])
    
#     if not second_call:
#         games_1718_x = games_1718_x.append(new_row)
#         games_1718_y = games_1718_y.append(row[["w_l"]])
#         second_call = True
#     else:
#         second_call = False

start_time = time.perf_counter()

games_1718.apply(create_vector, axis=1) # Replace w/ "sample" for debugging purposes

end_time = time.perf_counter()
print(f"Time to compute results: {Fore.BLUE} {end_time - start_time :.5f} {Fore.RESET} seconds")

print(games_1718_x.shape)
games_1718_x.head()

games_1718_y.head()

# Appending win loss values to training table.
combined = games_1718_x.copy()
combined.reset_index(drop=True, inplace=True)
combined['w_l'] = games_1718_y['w_l']
print(combined.shape)
combined.head()

#### Part 4: Exporting Data to be used for Training / Testing in Section 3

combined.to_csv("data/train_data.csv", index=False)
print(combined.columns)
