==> part4-model-construction.ipynb <==
## NCAA March Madness Prediction (Pt. 4)

import numpy as np
import pandas as pd
import sklearn
import sklearn.datasets
import sklearn.linear_model
import time
from colorama import Fore

import warnings
warnings.filterwarnings("ignore")

#### Part 1: Import Training Data

data = pd.read_csv("data/train_data.csv")
print("Shape:",data.shape)
print("Columns:", data.columns.values)

data.head(3)

X = data[data.columns.difference(["w_l"])]
X.head(3)

Y = (data.w_l == "W") # Defining 1 as Win, 0 as Loss
Y.head(3)

#### Part 2: Train / Test / Split Data

from sklearn.model_selection import train_test_split
x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.25, random_state=42)

#### Part 3: Logistic Regression Model 

# Defining Loss Function
def rmse(errors):
    return np.sqrt(np.mean(errors**2))

model = sklearn.linear_model.LogisticRegression(fit_intercept=True)
model.fit(x_train, y_train)

train_predict = model.predict(x_train)
test_predict = model.predict(x_test)

type(x_train)

train_accuracy = sum(train_predict == y_train) / len(y_train)
test_accuracy = sum(test_predict == y_test) / len(y_test)

# Uh oh, predictions are not very good, literally no better than flipping a coin
print("Training: ", train_accuracy)
print("Testing: ", test_accuracy)
